Previous commit broke resequencing if float values were given for a sequence.  This fixes that so float values can again be specified.
git-svn-id: http://dev.zenoss.com/svnint/trunk/core@16984 f5df01ce-0320-0410-9ed9-ddcd13c732cc

+++ Products/ZenModel/RRDTemplate.py
@@ -281,7 +281,7 @@ class RRDTemplate(ZenModelRM):
         if seqmap and origseq:
             try:
                 origseq = tuple([long(s) for s in origseq])
-                seqmap = tuple([long(s) for s in seqmap])
+                seqmap = tuple([float(s) for s in seqmap])
             except ValueError:
                 origseq = ()
                 seqmap = ()
